Убрана лишняя зависимость для обновления заголовка окна (#158)
