+++ package.json
@@ -50,7 +50,6 @@
     "monaco-editor-webpack-plugin": "^7.1.0",
     "nanoid": "^4.0.2",
     "re-resizable": "^6.9.11",
-    "react-document-title": "^2.0.3",
     "react-dropzone": "^14.2.3",
     "react-hook-form": "^7.45.1",
     "react-modal": "^3.16.1",
@@ -69,7 +68,6 @@
     "@types/lodash.throttle": "^4.1.8",
     "@types/node": "^18.16.16",
     "@types/react": "^18.2.8",
-    "@types/react-document-title": "^2.0.5",
     "@types/react-dom": "^18.2.4",
     "@types/react-modal": "^3.16.0",
     "@typescript-eslint/eslint-plugin": "^5.59.8",
